requirements.txt
langchain
langchain-google-genai
fastapi
uvicorn
python-dotenv
pydantic
httpx
